+++ client/cards/src/components/resetPassword/stepThree/ResetPasswordStepThree.jsx
@@ -4,6 +4,7 @@ import { useForm } from '../../../hooks/useForm';
 import { AuthContext } from '../../../contexts/AuthContext';
 
 import { NewPasswordInputs } from './newPasswordInputs/NewPasswordInputs';
+import { SubmitInputs } from './submitInputs/SubmitInputs';
 
 import background from '../../img/background.png';
 import styles from './ResetPasswordStepThree.module.css';
@@ -19,20 +20,10 @@ export const ResetPasswordStepThree = () => {
         confirmNewPassword: '',
     }, onResetPasswordSubmitStepThree)
 
-    const checkForEmptyValues = () => {
-        if (values.newPassword === '' || values.confirmNewPassword === '') {
-            return false
-        } else {
-            return true
-        }
-    }
-
     return (
 
         <div className={styles.loginContainer} style={{ backgroundImage: `url(${background})` }}>
-
             <div className={styles.innerContiner}>
-
                 <div className={styles.mainLoginField}>
                     <h1>Create New Password</h1>
                     <div className={styles.innerColumn}>
@@ -40,31 +31,18 @@ export const ResetPasswordStepThree = () => {
                             <form id="login" method='POST' onSubmit={onSubmit}>
                                 <div className={styles.fieldLogin}>
 
-                                    <NewPasswordInputs values={values} changeHandler={changeHandler}/>
+                                    <NewPasswordInputs values={values} changeHandler={changeHandler} />
 
                                     {errorResetPassword.error &&
                                         <p className={styles.errorField}>{errorResetPassword.error}</p>
                                     }
 
-                                    {checkForEmptyValues() &&
-                                        <div className="submit">
-                                            <input type="submit" className={styles.submitBtn} value="SUBMIT" />
-                                        </div>
-                                        ||
-                                        <div className="submit">
-                                            <input
-                                                type="submit"
-                                                className={styles.disabledSubmitBtn}
-                                                disabled={true}
-                                                title='Please Fill All Fields'
-                                                value="SUBMIT" />
-                                        </div>
-                                    }
+                                    <SubmitInputs values={values} />
+
                                 </div>
                             </form>
                         </section>
                     </div>
-
                 </div>
             </div>
         </div>
