+++ client/cards/src/components/resetPassword/stepThree/ResetPasswordStepThree.module.css
@@ -52,33 +52,6 @@
     background-color: #1E1F22;
 }
 
-.submitBtn {
-    width: 30%;
-    height: 30px;
-    margin-top: 20px;
-    background-color: #5865F2;
-    color: #F2F3F5;
-    border-radius: 5px;
-    border: none;
-    cursor: pointer;
-    transition: 0.3s;
-}
-
-.submitBtn:hover {
-    background-color: #3d46ae;
-}
-
-.loginContainer .disabledSubmitBtn {
-    width: 30%;
-    height: 30px;
-    margin-top: 20px;
-    background-color: #7b7d91;
-    color: #F2F3F5;
-    border-radius: 5px;
-    border: none;
-    cursor: no-drop;
-}
-
 .loginContainer .errorField {
     position: absolute;
     color: red;
