+++ client/cards/src/components/resetPassword/stepThree/submitInputs/SubmitInputs.module.css
@@ -1,0 +1,26 @@
+.submitBtn {
+    width: 30%;
+    height: 30px;
+    margin-top: 20px;
+    background-color: #5865F2;
+    color: #F2F3F5;
+    border-radius: 5px;
+    border: none;
+    cursor: pointer;
+    transition: 0.3s;
+}
+
+.submitBtn:hover {
+    background-color: #3d46ae;
+}
+
+.disabledSubmitBtn {
+    width: 30%;
+    height: 30px;
+    margin-top: 20px;
+    background-color: #7b7d91;
+    color: #F2F3F5;
+    border-radius: 5px;
+    border: none;
+    cursor: no-drop;
+}
